Moved JsTrait methods to ElementTrait.
