+++ tests/behat/bootstrap/FeatureContext.php
@@ -20,7 +20,6 @@ use DrevOps\BehatSteps\EmailTrait;
 use DrevOps\BehatSteps\FieldTrait;
 use DrevOps\BehatSteps\FileDownloadTrait;
 use DrevOps\BehatSteps\FileTrait;
-use DrevOps\BehatSteps\JsTrait;
 use DrevOps\BehatSteps\KeyboardTrait;
 use DrevOps\BehatSteps\LinkTrait;
 use DrevOps\BehatSteps\MediaTrait;
@@ -60,7 +59,6 @@ class FeatureContext extends DrupalContext {
   use FieldTrait;
   use FileDownloadTrait;
   use FileTrait;
-  use JsTrait;
   use KeyboardTrait;
   use LinkTrait;
   use MediaTrait;
